refactor: PositiveNumber VO를 통해서 숫자에 대한 책임을 부여
