+++ src/main/java/calculator/Input.java
@@ -1,5 +1,8 @@
 package calculator;
 
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
@@ -7,14 +10,15 @@ public class Input {
   private static final String SEPERATOR = "[,:]";
   private static final Pattern CUSTOM_SEPERATOR_PATTERN = Pattern.compile("//(.)\n(.*)");
 
-  private final String[] values;
+  private final List<PositiveNumber> positiveNumbers = new ArrayList<>();
 
   public Input(String inputString) {
-    this.values = split(inputString);
+    String[] values = split(inputString);
+    Arrays.stream(values).toList().forEach(value -> positiveNumbers.add(new PositiveNumber(value)));
   }
 
-  public String[] getValues() {
-    return values;
+  public List<PositiveNumber> getPositiveNumbers() {
+    return positiveNumbers;
   }
 
   private String[] split(String input) {
